refactor: extract project chat routes
