+++ src/server.js
@@ -14,7 +14,6 @@ import { execSync } from 'child_process';
 import yaml from 'js-yaml';
 import Database from 'better-sqlite3';
 import { runAgentWithAPI } from './agent-runner.js';
-import { listSessions as chatListSessions, createSession as chatCreateSession, getSession as chatGetSession, deleteSession as chatDeleteSession, updateSessionPreferences as chatUpdateSessionPreferences, streamChatMessage, getActiveStream, isStreaming as isChatStreaming, saveMessage as chatSaveMessage } from './chat.js';
 import { resolveModel, callModel, buildUserMessage, getModels as getPiModels } from './providers/index.js';
 import { buildCustomTierMap, resolveProviderRuntime } from './providers/custom-config.js';
 import { startOAuthLogin, submitManualCode, checkOAuthStatus, getAccessToken as getOAuthAccessToken, clearCredentials as clearOAuthCredentials, listOAuthProviders, loadCredentials as loadOAuthCredentials } from './oauth.js';
@@ -41,6 +40,7 @@ import { handleProjectIssueRoutes } from './server/routes/project/issues.js';
 import { handleProjectActionRoutes } from './server/routes/project/actions.js';
 import { handleProjectReportRoutes } from './server/routes/project/reports.js';
 import { handleProjectRepoRoutes } from './server/routes/project/repo.js';
+import { handleProjectChatRoutes } from './server/routes/project/chats.js';
 import webpush from 'web-push';
 import { config as loadDotenv } from 'dotenv';
 
@@ -3105,6 +3105,9 @@ const server = http.createServer(async (req, res) => {
       getOAuthAccessToken,
       getProviderRuntimeSelection,
       parseSummarizeCooldown,
+      parseExplicitModelSelection,
+      detectProviderFromToken,
+      formatStoredChatErrorMessage,
     };
 
     if (await handleProjectStatusRoutes(req, res, url, projectRouteContext)) return;
@@ -3119,407 +3122,7 @@ const server = http.createServer(async (req, res) => {
 
     if (await handleProjectRepoRoutes(req, res, url, projectRouteContext)) return;
 
-    // --- Chat endpoints ---
-
-    // GET /api/projects/:id/chats — list chat sessions
-    if (req.method === 'GET' && subPath === 'chats') {
-      try {
-        const sessions = chatListSessions(runner.chatsDir);
-        res.writeHead(200, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ sessions }));
-      } catch (e) {
-        res.writeHead(500, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: e.message }));
-      }
-      return;
-    }
-
-    // POST /api/projects/:id/chats — create new session
-    if (req.method === 'POST' && subPath === 'chats') {
-      if (!requireWrite(req, res)) return;
-      let body = '';
-      req.on('data', chunk => body += chunk);
-      req.on('end', () => {
-        try {
-          const data = body ? JSON.parse(body) : {};
-          const session = chatCreateSession(runner.chatsDir, data.title, {
-            selectedKeyId: typeof data.selectedKeyId === 'string' ? data.selectedKeyId : null,
-            selectedModel: typeof data.selectedModel === 'string' ? data.selectedModel : null,
-          });
-          res.writeHead(201, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ session }));
-        } catch (e) {
-          res.writeHead(500, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // GET /api/projects/:id/chats/:chatId — get session with messages
-    const chatDetailMatch = req.method === 'GET' && subPath.match(/^chats\/(\d+)$/);
-    if (chatDetailMatch) {
-      try {
-        const session = chatGetSession(runner.chatsDir, parseInt(chatDetailMatch[1]));
-        if (!session) {
-          res.writeHead(404, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: 'Session not found' }));
-        } else {
-          const chatId = parseInt(chatDetailMatch[1]);
-          const activeStream = getActiveStream(chatId);
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({
-            session,
-            streaming: !!activeStream,
-            streamingContent: activeStream ? { text: activeStream.text, toolCalls: activeStream.toolCalls } : null,
-          }));
-        }
-      } catch (e) {
-        res.writeHead(500, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: e.message }));
-      }
-      return;
-    }
-
-    // PATCH /api/projects/:id/chats/:chatId/preferences — persist session key/model selection
-    const chatPreferencesMatch = req.method === 'PATCH' && subPath.match(/^chats\/(\d+)\/preferences$/);
-    if (chatPreferencesMatch) {
-      if (!requireWrite(req, res)) return;
-      let body = '';
-      req.on('data', chunk => body += chunk);
-      req.on('end', () => {
-        try {
-          const data = body ? JSON.parse(body) : {};
-          const chatId = parseInt(chatPreferencesMatch[1]);
-          const session = chatUpdateSessionPreferences(runner.chatsDir, chatId, {
-            selectedKeyId: typeof data.selectedKeyId === 'string' && data.selectedKeyId !== 'auto' ? data.selectedKeyId : null,
-            selectedModel: typeof data.selectedModel === 'string' && data.selectedModel !== 'auto' ? data.selectedModel : null,
-          });
-          if (!session) {
-            res.writeHead(404, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify({ error: 'Session not found' }));
-            return;
-          }
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ session }));
-        } catch (e) {
-          res.writeHead(500, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // GET /api/projects/:id/chats/:chatId/stream — reconnect to active SSE stream
-    const chatStreamMatch = req.method === 'GET' && subPath.match(/^chats\/(\d+)\/stream$/);
-    if (chatStreamMatch) {
-      const chatId = parseInt(chatStreamMatch[1]);
-      const activeStream = getActiveStream(chatId);
-      if (!activeStream) {
-        res.writeHead(200, { 'Content-Type': 'text/event-stream', 'Cache-Control': 'no-cache', 'Connection': 'keep-alive' });
-        res.write(`data: ${JSON.stringify({ type: 'done' })}\n\n`);
-        res.end();
-        return;
-      }
-      res.writeHead(200, { 'Content-Type': 'text/event-stream', 'Cache-Control': 'no-cache', 'Connection': 'keep-alive' });
-      activeStream.clients.add(res);
-      res.on('close', () => { activeStream.clients.delete(res); });
-      return;
-    }
-
-    // DELETE /api/projects/:id/chats/:chatId — delete session
-    const chatDeleteMatch = req.method === 'DELETE' && subPath.match(/^chats\/(\d+)$/);
-    if (chatDeleteMatch) {
-      if (!requireWrite(req, res)) return;
-      try {
-        chatDeleteSession(runner.chatsDir, parseInt(chatDeleteMatch[1]));
-        res.writeHead(200, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ success: true }));
-      } catch (e) {
-        res.writeHead(500, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: e.message }));
-      }
-      return;
-    }
-
-    // POST /api/projects/:id/chats/upload — upload file for chat
-    if (req.method === 'POST' && subPath === 'chats/upload') {
-      if (!requireWrite(req, res)) return;
-      const chunks = [];
-      req.on('data', chunk => chunks.push(chunk));
-      req.on('end', () => {
-        try {
-          const body = Buffer.concat(chunks);
-          // Parse multipart boundary
-          const contentType = req.headers['content-type'] || '';
-          const boundaryMatch = contentType.match(/boundary=(.+)/);
-          if (!boundaryMatch) {
-            res.writeHead(400, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify({ error: 'Missing multipart boundary' }));
-            return;
-          }
-          const boundary = '--' + boundaryMatch[1];
-          const parts = body.toString('binary').split(boundary).filter(p => p.trim() && p.trim() !== '--');
-          
-          let filename = null;
-          let fileData = null;
-          let mimeType = null;
-
-          for (const part of parts) {
-            const headerEnd = part.indexOf('\r\n\r\n');
-            if (headerEnd === -1) continue;
-            const headers = part.slice(0, headerEnd);
-            const filenameMatch = headers.match(/filename="([^"]+)"/);
-            const ctMatch = headers.match(/Content-Type:\s*(.+)/i);
-            if (filenameMatch) {
-              filename = filenameMatch[1];
-              mimeType = ctMatch ? ctMatch[1].trim() : 'application/octet-stream';
-              // Extract binary data (skip headers + \r\n\r\n, remove trailing \r\n)
-              const dataStart = headerEnd + 4;
-              const dataEnd = part.endsWith('\r\n') ? part.length - 2 : part.length;
-              fileData = Buffer.from(part.slice(dataStart, dataEnd), 'binary');
-            }
-          }
-
-          if (!filename || !fileData) {
-            res.writeHead(400, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify({ error: 'No file in upload' }));
-            return;
-          }
-
-          // Save to uploads/
-          const uploadsDir = runner.uploadsDir;
-          if (!fs.existsSync(uploadsDir)) fs.mkdirSync(uploadsDir, { recursive: true });
-          const ext = path.extname(filename) || '.bin';
-          const safeName = `${Date.now()}-${Math.random().toString(36).slice(2, 8)}${ext}`;
-          fs.writeFileSync(path.join(uploadsDir, safeName), fileData);
-
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({
-            filename: safeName,
-            originalName: filename,
-            mimeType,
-            size: fileData.length,
-            url: `/api/projects/${projectId}/uploads/${safeName}`,
-          }));
-        } catch (e) {
-          res.writeHead(500, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // GET /api/projects/:id/uploads/:filename — serve uploaded files
-    const uploadMatch = req.method === 'GET' && subPath.match(/^uploads\/(.+)$/);
-    if (uploadMatch) {
-      const filename = uploadMatch[1];
-      const filePath = path.join(runner.uploadsDir, filename);
-      if (!fs.existsSync(filePath)) {
-        res.writeHead(404);
-        res.end('Not found');
-        return;
-      }
-      const ext = path.extname(filename).toLowerCase();
-      const mimeTypes = { '.jpg': 'image/jpeg', '.jpeg': 'image/jpeg', '.png': 'image/png', '.gif': 'image/gif', '.webp': 'image/webp', '.pdf': 'application/pdf', '.txt': 'text/plain', '.csv': 'text/csv' };
-      res.writeHead(200, { 'Content-Type': mimeTypes[ext] || 'application/octet-stream' });
-      fs.createReadStream(filePath).pipe(res);
-      return;
-    }
-
-    // POST /api/projects/:id/chats/:chatId/message — send message (SSE streaming)
-    const chatMessageMatch = req.method === 'POST' && subPath.match(/^chats\/(\d+)\/message$/);
-    if (chatMessageMatch) {
-      if (!requireWrite(req, res)) return;
-      const chatId = parseInt(chatMessageMatch[1]);
-      let body = '';
-      req.on('data', chunk => body += chunk);
-      req.on('end', async () => {
-        const respondChatError = (statusCode, payload) => {
-          try {
-            chatSaveMessage(runner.chatsDir, chatId, 'assistant', formatStoredChatErrorMessage({
-              error: payload.error,
-              statusCode,
-              source: payload.source || 'server',
-              cooldownMs: payload.cooldownMs || 0,
-            }), null, { success: false });
-          } catch {}
-          res.writeHead(statusCode, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify(payload));
-        };
-        try {
-          const data = JSON.parse(body);
-          if (!data.message?.trim()) {
-            respondChatError(400, { error: 'Message is required', source: 'validation' });
-            return;
-          }
-
-          // Save user message once for this send attempt
-          const imageUrls = (data.images || []).map(img => `/api/projects/${projectId}/uploads/${img.filename}`);
-          chatSaveMessage(runner.chatsDir, chatId, 'user', data.message.trim(), imageUrls.length > 0 ? imageUrls : null);
-
-          // Resolve API key / model selection
-          const config = runner.loadConfig();
-          const oauthTokenGetter = async (authFile, provider) => {
-            return getOAuthAccessToken(provider, runner.id);
-          };
-          const explicitKeyId = typeof data.keyId === 'string' && data.keyId.trim() ? data.keyId.trim() : null;
-          const { model: explicitModel, reasoningEffort: explicitReasoningEffort } = parseExplicitModelSelection(data.model);
-          if (explicitModel && !explicitKeyId) {
-            respondChatError(400, { error: 'Select a key before selecting a specific model.', source: 'validation' });
-            return;
-          }
-
-          const selectedKeySafe = explicitKeyId
-            ? (getKeyPoolSafe().keys || []).find(key => key.id === explicitKeyId) || null
-            : null;
-
-          chatUpdateSessionPreferences(runner.chatsDir, chatId, {
-            selectedKeyId: explicitKeyId,
-            selectedModel: explicitModel,
-          });
-          if (explicitKeyId && !selectedKeySafe) {
-            respondChatError(404, { error: 'Selected API key was not found.', errorType: 'key_not_found', source: 'local_selection' });
-            return;
-          }
-          if (selectedKeySafe && !selectedKeySafe.enabled) {
-            respondChatError(400, { error: 'Selected API key is disabled.', errorType: 'key_disabled', source: 'local_selection' });
-            return;
-          }
-          if (selectedKeySafe?.rateLimited) {
-            respondChatError(429, {
-              error: `Selected API key is currently rate limited${selectedKeySafe.cooldownMs ? ` for about ${Math.ceil(selectedKeySafe.cooldownMs / 60_000)}m` : ''}.`,
-              errorType: 'key_rate_limited',
-              source: 'local_cooldown',
-              cooldownMs: selectedKeySafe.cooldownMs || 0,
-            });
-            return;
-          }
-
-          const keyConfig = explicitKeyId
-            ? { ...config, keySelection: { keyId: explicitKeyId, fallback: false } }
-            : config;
-          const keyResult = await resolveKeyForProject(keyConfig, null, oauthTokenGetter);
-          if (!keyResult?.token) {
-            respondChatError(400, { error: explicitKeyId ? 'Selected API key is unavailable.' : 'No API key configured. Add one in Settings > Credentials.', source: explicitKeyId ? 'local_selection' : 'configuration' });
-            return;
-          }
-          if (explicitKeyId && keyResult.keyId !== explicitKeyId) {
-            respondChatError(400, { error: 'Selected API key is unavailable.', errorType: 'key_unavailable', source: 'local_selection' });
-            return;
-          }
-
-          const modelTier = data.modelTier || 'high';
-          const providerHint = keyResult.provider || detectProviderFromToken(keyResult.token);
-          const runtimeSelection = explicitModel
-            ? {
-                selectedModel: explicitModel,
-                reasoningEffort: explicitReasoningEffort,
-                customConfig: keyResult.customConfig || null,
-              }
-            : getProviderRuntimeSelection({
-                provider: providerHint,
-                modelTier,
-                keyResult,
-                projectModels: config.models,
-              });
-
-          // SSE headers
-          res.writeHead(200, {
-            'Content-Type': 'text/event-stream',
-            'Cache-Control': 'no-cache',
-            'Connection': 'keep-alive',
-          });
-
-          const chatOpts = {
-            agentDir: runner.chatsDir,
-            tbcDbPath: runner.projectDbPath,
-            uploadsDir: runner.uploadsDir,
-            projectPath: runner.path,
-            chatId,
-            userMessage: data.message.trim(),
-            images: data.images || [],
-            model: runtimeSelection.selectedModel,
-            token: keyResult.token,
-            provider: providerHint,
-            customConfig: runtimeSelection.customConfig || null,
-            res,
-            reasoningEffort: runtimeSelection.reasoningEffort || null,
-          };
-
-          try {
-            await streamChatMessage(chatOpts);
-          } catch (chatErr) {
-            // Check if rate-limited — try fallback key only in auto-key mode
-            const isRateLimit = /rate.limit|usage.limit|quota|429/i.test(chatErr.message);
-            if (isRateLimit && keyResult.keyId) {
-              const cooldownMs = parseSummarizeCooldown(chatErr.message);
-              markRateLimited(keyResult.keyId, cooldownMs);
-              log(`Chat: marked key ${keyResult.keyId} rate-limited for ${Math.ceil(cooldownMs / 60_000)}m`, runner.id);
-
-              if (explicitKeyId) {
-                chatErr.errorType = 'provider_rate_limited';
-                chatErr.source = 'provider_429';
-                chatErr.cooldownMs = cooldownMs;
-                chatErr.statusCode = 429;
-                throw chatErr;
-              }
-
-              // Try fallback key
-              const fallbackKey = await resolveKeyForProject(config, null, oauthTokenGetter);
-              if (fallbackKey?.token && fallbackKey.token !== keyResult.token) {
-                const fbProvider = fallbackKey.provider || detectProviderFromToken(fallbackKey.token);
-                const fallbackSelection = getProviderRuntimeSelection({
-                  provider: fbProvider,
-                  modelTier,
-                  keyResult: fallbackKey,
-                  projectModels: null,
-                });
-                log(`Chat: falling back to key ${fallbackKey.keyId} (${fbProvider}), model → ${fallbackSelection.selectedModel}`, runner.id);
-                chatOpts.token = fallbackKey.token;
-                chatOpts.provider = fbProvider;
-                chatOpts.model = fallbackSelection.selectedModel;
-                chatOpts.reasoningEffort = fallbackSelection.reasoningEffort || null;
-                chatOpts.customConfig = fallbackSelection.customConfig || null;
-                await streamChatMessage(chatOpts);
-              } else {
-                throw chatErr; // no fallback available
-              }
-            } else {
-              throw chatErr;
-            }
-          }
-
-          res.end();
-        } catch (e) {
-          const errorPayload = {
-            error: e.message,
-            errorType: e.errorType || null,
-            source: e.source || 'server',
-            statusCode: e.statusCode || 500,
-            cooldownMs: e.cooldownMs || 0,
-          };
-          try {
-            chatSaveMessage(runner.chatsDir, chatId, 'assistant', formatStoredChatErrorMessage(errorPayload), null, { success: false });
-          } catch {}
-          if (!res.headersSent) {
-            res.writeHead(errorPayload.statusCode, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify(errorPayload));
-          } else {
-            res.write(`data: ${JSON.stringify({
-              type: 'error',
-              content: errorPayload.error,
-              errorType: errorPayload.errorType,
-              source: errorPayload.source,
-              statusCode: errorPayload.statusCode,
-              cooldownMs: errorPayload.cooldownMs,
-            })}\n\n`);
-            res.end();
-          }
-        }
-      });
-      return;
-    }
+    if (await handleProjectChatRoutes(req, res, url, projectRouteContext)) return;
 
     if (await handleProjectActionRoutes(req, res, url, projectRouteContext)) return;
   }
